requirements.txt
playwright
streamlit
pandas
sqlalchemy
python-dotenv
beautifulsoup4
html5lib
psycopg2-binary
